+++ questions.js
@@ -1,10 +1,5 @@
 questions = [
     {
-        "type": "2-choice",
-        "question": "Does Connor love Paula?",
-        "correct": "True"
-    },
-    {
         "type": "4-choice",
         "question": "When did Connor and Paula start dating?",
         "choice1": "July 24, 2019",
@@ -12,5 +7,316 @@ questions = [
         "choice3": "June 24, 2020",
         "choice4": "June 19, 2020",
         "correct": "June 24, 2019"
+    },
+    {
+        "type": "4-choice",
+        "question": "When Pandora charm did Connor first buy for Paula?",
+        "choice1": "Lilo & Stitch Ohana Charm",
+        "choice2": "Mickey's Sorcerers Hat Charm",
+        "choice3": "Tower of Terror Charm",
+        "choice4": "Cinderella's Castle Charm",
+        "correct": "Lilo &amp; Stitch Ohana Charm"
+    },
+    {
+        "type": "4-choice",
+        "question": "Where did Paula first get to touch snow?",
+        "choice1": "Biltmore",
+        "choice2": "Blowing Rock",
+        "choice3": "A random street",
+        "choice4": "Appalachian State University",
+        "correct": "A random street"
+    },
+    {
+        "type": "4-choice",
+        "question": "What colors were the octopus Connor bought for Paula that never arrived?",
+        "choice1": "Blue & Red",
+        "choice2": "Blue & Yellow",
+        "choice3": "Black & White",
+        "choice4": "Purple & Green",
+        "correct": "Purple &amp; Green"
+    },
+    {
+        "type": "4-choice",
+        "question": "What move did the phrase 'trusting naive Paulie Girl' come from?",
+        "choice1": "Lilo & Stitch 2",
+        "choice2": "Lilo & Stitch: The Movie",
+        "choice3": "Lilo & Stitch",
+        "choice4": "Leroy & Stitch",
+        "correct": "Lilo &amp; Stitch 2"
+    },
+    {
+        "type": "4-choice",
+        "question": "Why didn't Connor want to meet Paula's friends at the poolhouse?",
+        "choice1": "Nerves",
+        "choice2": "Needed to use the bathroom",
+        "choice3": "Introversion",
+        "choice4": "All of these",
+        "correct": "All of these"
+    },
+    {
+        "type": "4-choice",
+        "question": "What meal in Mexico always gives Connor hot, boiling, goosebump-driven acidic diarrhea?",
+        "choice1": "Lonche",
+        "choice2": "Choripán",
+        "choice3": "Burger Club burger",
+        "choice4": "Carne Asada",
+        "correct": "Lonche"
+    },
+    {
+        "type": "4-choice",
+        "question": "What scene from The Clone Wars did Connor cry?",
+        "choice1": "When Vader takes Ahsoka's lightsaber",
+        "choice2": "The last time Ahsoka & Anakin see each other",
+        "choice3": "When Ahsoka & Rex bury the turned clones",
+        "choice4": "When Ahsoka removes Rex's helmet to see him crying",
+        "correct": "The last time Ahsoka &amp; Anakin see each other"
+    },
+    {
+        "type": "4-choice",
+        "question": "What French movie did Paula show Connor?",
+        "choice1": "La Haine",
+        "choice2": "Les Choristes",
+        "choice3": "La Maternelle",
+        "choice4": "Letter from Siberia",
+        "correct": "Les Choristes"
+    },
+    {
+        "type": "4-choice",
+        "question": "For what occasion did Connor make a surprise visit to Mexico?",
+        "choice1": "Christmas",
+        "choice2": "Paula's birthday",
+        "choice3": "Connor's birthday",
+        "choice4": "No occasion",
+        "correct": "Christmas"
+    },
+    {
+        "type": "3-choice",
+        "question": "Where were the sand dunes where Paula flew a kite?",
+        "choice1": "Jefferson State Park",
+        "choice2": "Jetton Park",
+        "choice3": "Jockey's Ridge State Park",
+        "correct": "Jockey's Ridge State Park"
+    },
+    {
+        "type": "3-choice",
+        "question": "How many shows has Connor watched on Paula's recommendation?",
+        "choice1": "1",
+        "choice2": "2",
+        "choice3": "3",
+        "correct": "3"
+    },
+    {
+        "type": "3-choice",
+        "question": "How many shows has Paula watched on Connor's recommendation?",
+        "choice1": "1",
+        "choice2": "2",
+        "choice3": "3",
+        "correct": "1"
+    },
+    {
+        "type": "3-choice",
+        "question": "Who makes the best tasting grilled cheese, but isn't Connor's favorite?",
+        "choice1": "Sonic",
+        "choice2": "Char-Grill",
+        "choice3": "Five Guys",
+        "correct": "Five Guys"
+    },
+    {
+        "type": "3-choice",
+        "question": "Where was the photo from Paula's playlist for Connor taken?",
+        "choice1": "Biltmore",
+        "choice2": "Blowing Rock",
+        "choice3": "Blue Ridge State Park",
+        "correct": "Blowing Rock"
+    },
+    {
+        "type": "3-choice",
+        "question": "What meal did Connor make for Paula's family?",
+        "choice1": "Breaded Macaroni & Cheese",
+        "choice2": "Chicken Parmesan",
+        "choice3": "Chicken Quesadilla",
+        "correct": "Breaded Macaroni &amp; Cheese"
+    },
+    {
+        "type": "3-choice",
+        "question": "What word in English does Paula always forget?",
+        "choice1": "Yawn",
+        "choice2": "Yak",
+        "choice3": "Yam",
+        "correct": "Yawn"
+    },
+    {
+        "type": "3-choice",
+        "question": "What is along the frame of the picture Paula gave to Connor?",
+        "choice1": "A solid blue frame",
+        "choice2": "Jetton Park",
+        "choice3": "Loteria cards",
+        "correct": "Loteria cards"
+    },
+    {
+        "type": "3-choice",
+        "question": "What game did Paula buy for Connor?",
+        "choice1": "Super Mario 3D All-Stars",
+        "choice2": "Super Paper Mario: The Origami King",
+        "choice3": "Super Mario Galaxy",
+        "correct": "Super Mario 3D All-Stars"
+    },
+    {
+        "type": "3-choice",
+        "question": "Which is the best of the Indiana Jones movies that matter?",
+        "choice1": "Raiders of the Lost Ark",
+        "choice2": "The Temple of Doom",
+        "choice3": "The Last Crusade",
+        "correct": "The Last Crusade"
+    },
+    {
+        "type": "2-choice",
+        "question": "Does Connor love Paula?",
+        "correct": "True"
+    },
+    {
+        "type": "2-choice",
+        "question": "Tú no eres para mí is Connor's Spanish guilty pleasure song?",
+        "correct": "True"
+    },
+    {
+        "type": "2-choice",
+        "question": "Does Paula love the song acid?",
+        "correct": "False"
+    },
+    {
+        "type": "2-choice",
+        "question": "Connor and Paula brought eggnog when going to see the Charlotte Motor Speedway Christmas lights",
+        "correct": "False"
+    },
+    {
+        "type": "2-choice",
+        "question": "Did Paula get Connor a bowling pin for his 22nd birthday?",
+        "correct": "True"
+    },
+    {
+        "type": "2-choice",
+        "question": "Does Connor like Paula's actual bed more than her side bed with a mattress topper?",
+        "correct": "False"
+    },
+    {
+        "type": "2-choice",
+        "question": "Paula has finished watching a movie with Connor without playing Gardenscapes",
+        "correct": "False"
+    },
+    {
+        "type": "2-choice",
+        "question": "Connor was put at the top of the first photo with Paula's family so he could be easily removed",
+        "correct": "True"
+    },
+    {
+        "type": "2-choice",
+        "question": "Connor said 'I love you' first",
+        "correct": "True"
+    },
+    {
+        "type": "2-choice",
+        "question": "Connor and Paula would sit in the car outside her apartment and talk until 1 or 2 AM",
+        "correct": "True"
+    },
+    {
+        "type": "image",
+        "image": "img/questions/alex.webp",
+        "question": "Who took this photo?",
+        "choice1": "Jonathan",
+        "choice2": "Chip",
+        "choice3": "Alex",
+        "choice4": "Aiden",
+        "correct": "Alex"
+    },
+    {
+        "type": "image",
+        "image": "img/questions/casting.webp",
+        "question": "Where is this location?",
+        "choice1": "Disney University",
+        "choice2": "Casting Office",
+        "choice3": "PhilharMagic",
+        "choice4": "Beauty & The Beast Show",
+        "correct": "Casting Office"
+    },
+    {
+        "type": "image",
+        "image": "img/questions/durham.webp",
+        "question": "Where was this picture taken?",
+        "choice1": "Raleigh",
+        "choice2": "Charlotte",
+        "choice3": "Kitty Hawk",
+        "choice4": "Durham",
+        "correct": "Durham"
+    },
+    {
+        "type": "image",
+        "image": "img/questions/garden.webp",
+        "question": "Where was this picture taken?",
+        "choice1": "Daniel Stowe Botanical Garden",
+        "choice2": "Biltmore",
+        "choice3": "Jetton Park",
+        "choice4": "Charlotte Motor Speedway",
+        "correct": "Daniel Stowe Botanical Garden"
+    },
+    {
+        "type": "image",
+        "image": "img/questions/lake.webp",
+        "question": "Where was this picture taken?",
+        "choice1": "Lago de Chapala",
+        "choice2": "Laguna de Cajititlán",
+        "choice3": "Jocotepec",
+        "choice4": "Laguna de Sayula",
+        "correct": "Lago de Chapala"
+    },
+    {
+        "type": "image",
+        "image": "img/questions/sinatra.webp",
+        "question": "What is the name of this dog from DCC?",
+        "choice1": "Tank",
+        "choice2": "Sinatra",
+        "choice3": "Scout",
+        "choice4": "Bryan",
+        "correct": "Sinatra"
+    },
+    {
+        "type": "image",
+        "image": "img/questions/discovery.webp",
+        "question": "Where was this photo taken?",
+        "choice1": "Discovery Place",
+        "choice2": "Sally's Y Trail",
+        "choice3": "Jetton Park",
+        "choice4": "Trail in Connor's Neighborhood",
+        "correct": "Discovery Place"
+    },
+    {
+        "type": "image",
+        "image": "img/questions/golf.webp",
+        "question": "What mini-golf course was this photo taken?",
+        "choice1": "Sunken Treasure Mini-Golf",
+        "choice2": "Mutiny Bay Adventure Golf",
+        "choice3": "Lost Treasure Golf",
+        "choice4": "Pirate's Adventure Mini-Golf",
+        "correct": "Mutiny Bay Adventure Golf"
+    },
+    {
+        "type": "image",
+        "image": "img/questions/hatteras.webp",
+        "question": "Where was this photo taken?",
+        "choice1": "Bodie Island Lighthouse",
+        "choice2": "Oak Island Lighthouse",
+        "choice3": "Ocracoke Lighthouse",
+        "choice4": "Cape Hatteras Lighthouse",
+        "correct": "Cape Hatteras Lighthouse"
+    },
+    {
+        "type": "image",
+        "image": "img/questions/jetton.webp",
+        "question": "When was this photo taken?",
+        "choice1": "May 22, 2021",
+        "choice2": "May 24, 2020",
+        "choice3": "May 24, 2021",
+        "choice4": "May 26, 2021",
+        "correct": "May 24, 2021"
     }
 ]